Move CaseInsensitiveName into Naming namespace
