+++ src/Provider/PhpBenchUsageProvider.php
@@ -16,6 +16,7 @@ use PHPStan\PhpDocParser\Parser\TokenIterator;
 use ShipMonk\PHPStan\DeadCode\Graph\ClassMethodRef;
 use ShipMonk\PHPStan\DeadCode\Graph\ClassMethodUsage;
 use ShipMonk\PHPStan\DeadCode\Graph\UsageOrigin;
+use ShipMonk\PHPStan\DeadCode\Naming\CaseInsensitiveName;
 use function array_merge;
 use function explode;
 use function is_array;
